Updated the file structure to include server.js for the variables that will import express, turn app into the express function, use the app.use method to implement middleware to the express application so that the application can recieve parsed JSON data to turn into objects which can come from the url or the mounted /api paths that will be implemented soon however I have not fixed yet, and must include the get method, post method, and listen method/Updated the file structure to include a routes folder for the notes.js which will be for getting the routes, reading, appending, writing to the files, and deleting soon however I have not implemented any of this stuff so will fix/Updated the files structure to include a helpers folder for the utilities which will include functions for reading from and writing to JSON files.
Updated the file structure to include server.js for the variables that will import express, turn app into the express function, use the app.use method to implement middleware to the express application so that the application can recieve parsed JSON data to turn into objects which can come from the url or the mounted /api paths that will be implemented soon however I have not fixed yet, and must include the get method, post method, and listen method. Updated the file structure to include a routes folder for the notes.js which will be for getting the routes, reading, appending, writing to the files, and deleting soon however I have not implemented any of this stuff so will fix. Updated the files structure to include a helpers folder for the utilities which will include functions for reading from and writing to JSON files.