Updated the fsUtils.js file to include a comment for the variables that require the methods fs, and util/Updated db.json to include a code snippet that will help test whether a a note has a title and notes in it/Updated the notes.js to include the variables for requiring the express module, the uuid module, and the helper functions from the fsUtils.js, added a get route that retrieves all of the notes by using the get method with the readFromFile method which reads from the argument passed which is the db the database then returns a promise, then the promise is read successfully the call back function is passed to the then method which then will take the data for the next function to handle the response in json as it gets parsed out, the next get method which takes the applications notes requests the end point /:note_id, then when this request is made to the endpoint the value note_id is extracted from the URL, then when it calls from the readFromFile function and the data is read successfully if there are objects that match the note_id the code will send a JSON response with a filtered result using res.json if no note_id then the JSON response send a message 'No note with that ID' using the else : syntax, then the notes will be grabbed using the delete method to delete a specific note from a list of notes when the request is made to the endpoint /:note_id so when it is deleted the code will read from a filtered array of notes to create a new array that excludes the note with the ID matching the noteId pulled from the request, then the Json response is sent back to the client indicating that the note has been deleted, then the code using the post method for the requests made to the endpoint /notes will log the body request to the console, then take the username, topic, and note properties however these arent the right values will fix, although when the values are grabbed the readAndAppend method appends a newNote object with a JSON response that states a note has been added successfully, then the notes.js file is modularized for import in other files.
Updated the fsUtils.js file to include a comment for the variables that require the methods fs, and util so then they are modularized for use later on. Updated db.json to include a code snippet that will help test whether a a note has a title and notes in it for ease of TDD. Updated the notes.js to include the variables for requiring the express module, the uuid module, and the helper functions from the fsUtils.js, added a get route that retrieves all of the notes by using the get method with the readFromFile method which reads from the argument passed which is the db the database then returns a promise, then the promise is read successfully the call back function is passed to the then method which then will take the data for the next function to handle the response in json as it gets parsed out, the next get method which takes the applications notes requests the end point /:note_id, then when this request is made to the endpoint the value note_id is extracted from the URL, then when it calls from the readFromFile function and the data is read successfully if there are objects that match the note_id the code will send a JSON response with a filtered result using res.json if no note_id then the JSON response send a message 'No note with that ID' using the else : syntax, then the notes will be grabbed using the delete method to delete a specific note from a list of notes when the request is made to the endpoint /:note_id so when it is deleted the code will read from a filtered array of notes to create a new array that excludes the note with the ID matching the noteId pulled from the request, then the Json response is sent back to the client indicating that the note has been deleted, then the code using the post method for the requests made to the endpoint /notes will log the body request to the console, then take the username, topic, and note properties however these arent the right values will fix, although when the values are grabbed the readAndAppend method appends a newNote object with a JSON response that states a note has been added successfully, then the notes.js file is modularized for import in other files.