Updated the file structure to add middleware/Updated the middleware folder to include clog.js/Updated clog.js to include a logging function which is modularized, this console logging function will log out the type and path of each request to the server so that I can use it to see what is happening when I use get, and post methods in my server/Updated server.js to include the variables for requiring express, path, the clog function, the notes.js routes, and then included the port so that the PORT variable can be called and the server can run on the local environment, and then included app.use(clog); that way when the methods are called through the server I can see the console information.
Updated the file structure to add middleware/Updated the middleware folder to include clog.js/Updated clog.js to include a logging function which is modularized, this console logging function will log out the type and path of each request to the server so that I can use it to see what is happening when I use get, and post methods in my server so its easier to work with. Updated server.js to include the variables for requiring express, path, the clog function, the notes.js routes, and then included the port so that the PORT variable can be called and the server can run on the local environment, and then included app.use(clog); that way when the methods are called through the server I can see the console information.