Updated notes.json to be db.json per say the grading requirements/Updated notes.js to include a title for the newNote that is appended to the database(db.json), and then also removed username, and topic from the if statement, then changed all of the instances of notes.json to db.json/Updated server.js to include middle ware for parsing JSON and urlencoded form data, then added the middleware for the /api endpoint however it is not connecting to any route at the moment will fix, then added middleware for serving the static files, added a get route for the homepage, a get route for the notes page, and a listener for the server so that when npm run start is executed the port can be entered into the url and the server can be viewed.
Updated notes.json to be db.json per say the grading requirements. Updated notes.js to include a title for the newNote that is appended to the database(db.json), and then also removed username, and topic from the if statement, then changed all of the instances of notes.json to db.json. Updated server.js to include middle ware for parsing JSON and urlencoded form data, then added the middleware for the /api endpoint however it is not connecting to any route at the moment will fix, then added middleware for serving the static files, added a get route for the homepage, a get route for the notes page, and a listener for the server so that when npm run start is executed the port can be entered into the url and the server can be viewed.