Updated the routes file to have an api.js folder for the routes to the index.html however its not functioning yet/Updated the notes.js to change the constant variable on line 49 to include the title, and the note values from the if statements new note on lines 53, and 54 so that when I click save note on the page it saves however the saving does not work will fix.
Updated the routes file to have an api.js folder for the routes to the index.html however its not functioning yet will fix. Updated the notes.js to change the constant variable on line 49 to include the title, and the note values from the if statements new note on lines 53, and 54 so that when I click save note on the page it saves however the saving does not work will fix.