Updated the notes.js to include on line 49 a title and note that will specifically check if title and note exist and are not empty strings, added a then to promise that if the note is added to the database then respond with 'note added successfully' then a catch right after that states there was an error adding a note else if only a title or a text is added then send out a response stating the title and note are both required/Updated server.js to have a wildcard page that sends you back to the index.html if there is an error.
Updated the notes.js to include on line 49 a title and note that will specifically check if title and note exist and are not empty strings, added a then to promise that if the note is added to the database then respond with 'note added successfully' then a catch right after that states there was an error adding a note else if only a title or a text is added then send out a response stating the title and note are both required. Updated server.js to have a wildcard page that sends you back to the index.html if there is an error.