Updated db.json to add on the id so that I can select the posted notes by the id for deletion however it does not work yet/Updated store.js with a comment not however there is no code yet/Updated notes.js to include the correct route for the GET method on line 11 so that when I add a note it appends to the left column, and also fixed the route for the deletion button/Update server.js to comment out the GET method on line 23 so that I can have a route for the api to get rendered notes however I couldnt get it to function will fix.
Updated db.json to add on the id so that I can select the posted notes by the id for deletion however it does not work yet will fix. Updated store.js with a comment not however there is no code yet will fix. Updated notes.js to include the correct route for the GET method on line 11 so that when I add a note it appends to the left column, and also fixed the route for the deletion button however the deletion button does not work yet. Updated server.js to comment out the GET method on line 23 so that I can have a route for the api to get rendered notes however I couldnt get it to function will fix.