Updated notes.js to change lines 57-58 so that the newNote will be added to the database json file, changed line 50 to state if title and text not title and note are existing in the values of the form, on line 36 instead of note.note_id its now note.id/Updated file structure to move everything out of the develop folder so it will deploy correctly with render.
Updated notes.js to change lines 57-58 so that the newNote will be added to the database json file, changed line 50 to state if title and text not title and note are existing in the values of the form, on line 36 instead of note.note_id its now note.id. Updated file structure to move everything out of the develop folder so it will deploy correctly with render.